WAITM-165 Split out generic version of printouts
